requirements.txt
crewai>=0.28.0
python-dotenv
pydantic